requirements.txt
cffi >= 0.8.0
letsencrypt >= 0.3.0
six>=1.7
pyyaml>=3.11
